Move RepeatedDuration to Utils package
