+++ sdk/src/test/java/io/dapr/utils/RepeatedDurationTest.java
@@ -1,5 +1,6 @@
-package io.dapr.actors;
+package io.dapr.utils;
 
+import io.dapr.utils.RepeatedDuration;
 import org.junit.Test;
 
 import java.time.Duration;
